+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Generate API clients using Handlebars templates, with an option to provide custom templates (#75)
 
+### Fixed
+
+- Strip trailing slash in API clients (#87)
+
 ## [0.12.7] - 2024-10-10
 
 ### Fixed